README.md
Bank Customer Turnover Classification using TensorFlow and Keras
This project aims to predict whether a customer will continue with the bank or not (churn) using machine learning. The dataset used in this project is taken from Kaggle’s “Bank Turnover Dataset” 1. The dataset contains 10,000 different data points with 14 features.

Requirements
Python 3.6 or higher
TensorFlow 2.0 or higher
Keras 2.3.1 or higher
Installation
Clone the repository.
Install the required packages using pip install -r requirements.txt.
Usage
Run the bank-customer-churn-ann.ipynb file to train and test the model.
Results
The model achieved an accuracy of 85% on the test set.

Conclusion
In this project, we used TensorFlow and Keras to build a machine learning model to predict whether a customer will continue with the bank or not. The model achieved an accuracy of 85% on the test set. This project can be further improved by using more advanced machine learning techniques and by collecting more data.